+++ src/main/java/be/softwarelab/dao/UsersDao.java
@@ -43,14 +43,14 @@ public class UsersDao {
     }
     
     // Retrieves all the guests:
-    public Long getMaxId() {
-        TypedQuery<Long> query = em.createQuery(
-            "SELECT MAX(g.id) FROM Users g", Long.class);
+    public Integer getMaxId() {
+        TypedQuery<Integer> query = em.createQuery(
+            "SELECT MAX(g.id) FROM Users g", Integer.class);
         if (query.getResultList() != null && query.getResultList().size() > 0)
         {
             return query.getResultList().get(0);
         } else {
-            return 1L;
+            return 1;
         }
     }
